requirements.txt
# ============================================================
# Aerial Object Classification & Detection - Requirements
# ============================================================

# Deep Learning Frameworks
tensorflow>=2.12.0
keras>=2.12.0

# YOLOv8 (Object Detection)
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0

# Data Science & ML
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Streamlit Deployment
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Model utilities
h5py>=3.8.0
tensorboard>=2.12.0

# Image processing extras
albumentations>=1.3.0
imgaug>=0.4.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.28.0
